query-to-plots/requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
jupyter==1.0.0
matplotlib==3.8.4
seaborn==0.13.2
notebook==7.1.2
langchain-community==0.3.24
sqlparse==0.5.3
rich==14.0.0
langchain-openai==0.3.18

# Use precompiled binary to avoid build issues
psycopg2-binary==2.9.10

plotly==6.1.2

# FastAPI + Uvicorn with standard extras (includes dependencies like pydantic, typing-extensions, etc.)
fastapi[standard]==0.115.1
uvicorn[standard]==0.34.2
kaleido==1.0.0rc13
